Fixed issue with cards only loading on page refresh
Had to set the token in the apiObj to null initially, and then set it to the value when the user logs in.
Which means I had to add a setToken() function to Api.js
Also messed with useEffect so that:
1.The getuserinfo one runs whenever the token changes
2.The cards loading one runs whenever the currentUser changes.
And I removed [history]- don't re run stuff on history, just terrible idea really.